Changed box_score plus_minus column to type FLOAT instead of INTEGER

+++ triple_triple_etl/load/postgres/nbastats_postgres_etl.py
@@ -72,7 +72,7 @@ def create_nbastats_tables():
           turnovers INTEGER,
           personal_fouls INTEGER,
           points INTEGER,
-          plus_minus INTEGER            
+          plus_minus FLOAT            
         );
         END TRANSACTION;
     """
